60222581_/개별과제_5/test/Recommendation/ClothingItem.java
public class ClothingItem {
    private String itemID;
    private String itemName;

    public ClothingItem(String itemID, String itemName) {
        this.itemID = itemID;
        this.itemName = itemName;
    }

    public String getItemID() {
        return itemID;
    }

    public void setItemID(String itemID) {
        this.itemID = itemID;
    }

    public String getItemName() {
        return itemName;
    }

    public void setItemName(String itemName) {
        this.itemName = itemName;
    }
}
